refactor(chat): unify header + sidebar add-participant into one dropdown
Both the chat header quick-add icon and the right-sidebar inline "Add"
row now render a shared <AddParticipantDropdown>. A `variant` prop
("icon" | "inline") swaps only the trigger shell; the candidate dropdown
is identical across both surfaces — avatar + "mine / others" grouping
with a divider — and the component owns disclosure (open / click-outside
/ Esc), keyboard navigation, and the immediate addMeChatParticipants
mutation.

Also drops the duplicate chat-detail + activity queries the participants
section was re-declaring: membership data (participants / managedByMe) is
now passed down from ChatView, which already holds those queries.

The new-chat ParticipantChips (deferred-commit chip model + @-mention
linkage) is intentionally left untouched.
